+++ R/broom.helpers-package.R
@@ -1,6 +1,6 @@
 ## usethis namespace: start
 #' @importFrom lifecycle deprecate_soft
-#' @importFrom usethis ui_code
+#' @importFrom cli cli_alert_info cli_alert_info cli_alert_danger cli_code cli_ul
 #' @importFrom rlang .data .env
 #' @importFrom purrr %||%
 ## usethis namespace: end