+++ R/tidy_identify_variables.R
@@ -92,14 +92,14 @@ tidy_identify_variables <- function(x, model = tidy_get_model(x),
       tidy_attach_model(model = model, .attributes = .attributes)
   } else {
     if (!quiet)
-      usethis::ui_oops(paste0(
+      cli_alert_danger(paste0(
         "Unable to identify the list of variables.\n\n",
-        "This is usually due to an error calling {usethis::ui_code('stats::model.frame(x)')}",
-        "or {usethis::ui_code('stats::model.matrix(x)')}.\n",
+        "This is usually due to an error calling {.code stats::model.frame(x)}",
+        "or {.code stats::model.matrix(x)}.\n",
         "It could be the case if that type of model does not implement these methods.\n",
         "Rarely, this error may occur if the model object was created within\na ",
-        "functional programming framework (e.g. using {usethis::ui_code('lappy()')}, ",
-        "{usethis::ui_code('purrr::map()')}, etc.)."
+        "functional programming framework (e.g. using {.code lappy()}, ",
+        "{.code purrr::map()}, etc.)."
       ))
     if (strict) stop("Cannot identify variables. Quitting execution.", call. = FALSE)