+++ NEWS.md
@@ -4,6 +4,7 @@
   `"ran_vals"` for random-effect parameters and values, based of the 
   `effect` column returned by `broom.mixed::tidy()` (#90)
 - Support of `brms::brm()` models (#89)
+- Basic support of `cmprsk::crr()` models (#91)
 - Models with categorical variable and no intercept now supported (#85)
 - New contrasts type ("no.contrast") returned by `model_list_contrasts`()
 - Added support for `mgcv::gam()` models. (#82)