+++ NEWS.md
@@ -1,13 +1,13 @@
 # broom.helpers (development version)
 
 - Replaced `usethis::ui_*()` messaging with `cli::cli_*()` (#94)
-
 - For mixed models, `var_type` column is now equal to `"ran_pars"` or 
   `"ran_vals"` for random-effect parameters and values, based of the 
   `effect` column returned by `broom.mixed::tidy()` (#90)
 - Support of `lfe::felm()` models (#79)
 - Support of `brms::brm()` models (#89)
 - Basic support of `cmprsk::crr()` models (#91)
+- `strict` argument removed from `tidy_identify_variables()` (#99)
 - Models with categorical variable and no intercept now supported (#85)
 - New contrasts type ("no.contrast") returned by `model_list_contrasts`()
 - Added support for `mgcv::gam()` models. (#82)