+++ NEWS.md
@@ -6,6 +6,7 @@
   contrasts (#108)
 - Now `tidy_add_n()` works with multinomial models when `y` is not coded as
   a factor (#109)
+- `glue` added to Suggests
 
 # broom.helpers 1.2.0